[components] Improve errors when routing from void
When routing from void, a `debug!()` log is logged to the kernel log,
and a "SourceInstanceNotFound" error is logged to the component's log
sink. This doesn't match other kind of availability-related errors:

- Logging to kernel is much less useful than the component's log sink.
- SourceInstanceNotFound is the wrong error

So we change it to a dedicated void source error and add a test.

Change-Id: I0fea78b63df696c424e95a26c0f6c313b4322fb5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/985552
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Yifei Teng <yifeit@google.com>