[components] Move start/lookup methods from Model to ComponentInstance
This is a prefactor for b/320698181 which needs access to EventRegistry
in the Resolve action.

The idea is to move EventRegistry into ModelContext, but that creates
a circular dependency because EventRegistry depends on Model and uses
these methods. In a following change, EventRegistry will depend only
on the root component.

Change-Id: I6fd4d1c32f38f9e1559ce51378297cce080849f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986244
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Yegor Pomortsev <ypomortsev@google.com>