[superproject] Roll llvm-project/libc [libc] Remove CPU dependent AMDGPU instructions (#80707)
Summary:
Some recent changes allowed us to remove target level divergence one
these instructions. This patch removes the wavefront dependent
divergence for the ballot and thread ID functions, as well as the clock.
The changes to the "Vendor" library simply disables target specific
optimizations in the implementation. This should be removed in its
entirety when the LLVM `libm` is sufficiently implemented.

The remaining areas of divergence is only the RPC packet size and the
fixed frequency counter.

GitOrigin-RevId: 2d66058678fc6993b2d8822ac124080d4655268b
Original-Revision: e2cad3814fbbc43c7c6ba4e026dfaba067f0094d
Roller-URL: https://ci.chromium.org/b/8756827974264293633
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I56db2e3d91a9c494b57ea04590c5e964c3380145
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986932