[bazel] use target_name for build-id dirs
The bazel SDK was writing the build_id directories to
package_name/.build_id but this was causing collisions when there were
multiple packages defined in the same BUILD.bazel file because they
would both try to write to this directory.

This change puts the build id directory in ctx.label.name
+ _build_id_dir which makes sure it is unique to the bazel package and
target. Additionally, we had to update the bazel_action to look for the
build id directories at the new location. The code was updated to use
scopes instead of strings so we could piggyback off of the name
templating that is already built into bazel_action.

BUG: b/322752815
Change-Id: I2839ad47f2544fbf5c16ceddfea43835e24a6f87
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986110
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: David Fang <fangism@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>