[debugger] Log output should respect OutputEnabled
When getting log notification requests, we should respect the
OutputEnabled bit that is checked in the `Output` path, but not the
`Write` path. When operating in embedded mode, we really don't want
spurious DebugAgent logs getting mixed in with the logs we're likely
streaming via fd_streamers.

This was particularly noticeable if you ctrl+c'd rust tests being run
under fx test with any of the debug options and then getting a bunch of
"Process with KOID not found" messages as DebugAgent was cleaning up.

Change-Id: I1fb4700489cdd43e6f7f22b30da24cd1d1ee1fa7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984586
Fuchsia-Auto-Submit: Jacob Rutherford <jruthe@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Adam Barth <abarth@google.com>