[logs] Make Severity a u8
Per the new design, make Severity a u8
instead of a FIDL type. The FIDL type
is still used for cases where a Severity
is bounded to a FIDL severity, and for
display purposes, but is no longer
constrained in the wire format itself.

Bug: b/322552971
Bug: b/300157652
Change-Id: Ic348ba57170df0427f0c6e7319fa217b57b1c360
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986950
Reviewed-by: Miguel Flores <miguelfrde@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Brian Bosak <bbosak@google.com>
API-Review: Miguel Flores <miguelfrde@google.com>