[ffx][fastboot]: add rediscover function
InterfaceFactoryBase now has function rediscover(&mut self) which is
intended to be used during the flashing process in order to rediscover
the target after it has been instructed to reboot to the bootloader.

This code is not wired up to be used yet.

The current implementation of the reconnect functions assume:
  1. A target will never change fastboot connection mechanisms. For
     example, it is assumed that if a target is connected via USB, after
     the bootloader reboots it will continue to be available via USB
  2. A target's name (for fastboot over TCP/UDP) is consistent across
     reboots.
  3. A target's serial number (for fastboot over USB) is consistent
     across reboots.

Change-Id: Ib6a0537803033a4d702a38b951acea124636b4ef
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/984588
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>