[ffx][bootloader]: wait in a loop for Target to Reboot
When first executing the bootloader command, before any fastboot code is
executed, we need to ensure the Target is rebooted to the bootloader.
Instead of waiting for a flat 10 seconds (brittle), wait in a loop.

If the target is not discovered in Bootloader mode in 20 seconds, warn the
user.

Change-Id: I54a4163e8327e271d5b5eee294e945a7df832720
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986232
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Colin Nelson <colnnelson@google.com>