[fxfs] Add layer files to debug directory
Exposes persistent layer files in the debug directory.  The in-memory
layer is not exposed for now.

This required adding a number of callbacks for volume addition/removal
and object store flushing so the debug directory could be updated, and
making the debug directory hold weak references back to the main Fxfs
structures to avoid cycles.

I also fixed store_info.txt to update dynamically when the store is
flushed.

Bug: 323571978
Change-Id: I50879f21dc428fc0110584bfecb6f18c6db00a01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986094
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>