[netstack3] Add optional local delivery device
...to ResolvedRoute which carries the device that the destination
address is assigned to in the local delivery case. This enables
`new_ip_socket` to bind the socket to this device if either the source
or destination addresses require a zone ID. Prior to this change,
the socket was being bound to loopback, which is incorrect.

Return an error from resolve_route if the source and destination
addresses are assigned to different interfaces, and one of them
requires a zone ID. Note that Linux does allow this, but exhibits
strange behavior. This case is difficult to support correctly, so
we prohibit it for now.

Skip the integration test that pings an IPv6 LL address assigned to
oneself, which used to fail due to a routing error, but now stalls
forever because delivery of the echo reply cannot make it back to
the ICMP socket due to fxbug.dev/42084713.

Fixed: 323958072
Bug: 42084713
Test: fx test --test-filter '*lookup_route*' netstack3-core-test
Change-Id: Ie943845d427955762d1a486774411110026cc2a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/982054
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>