[build] Disable -Wextra-qualification
This works around an issue with -Wextra-qualification, where fidl
generated code is causing this warning to fire and break the build with
newer versions of clang. It is currently unclear what the underlying
issue is, but this flag can be removed once a long term fix is available
in the compiler or in the Fuchsia tree.

Note: -Wextra-qualification warns about unneeded qualifiers on members.

Bug: 324114504
Change-Id: I848cc2a03db00df76da1cb9abccd6fa965042c9f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987955
Commit-Queue: Paul Kirth <paulkirth@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>