[fx][serve] Warn about running foreground server not being detected
When the ffx foreground repo server (currently in beta) is enabled (by
setting the environment `FUCHSIA_DISABLED_foreground_repo_server=0`),
it is not currently not possible for `fx serve` to detect if an ffx
foreground repo server instance is running, or if something else is
merely using the same port. Warn the user about this, and suggest
the workaround to disable the ffx repo server listening on a port.

Fixed: b/323583171
Change-Id: I76d6c06abda66079da48531425e3d0619f0f9d4f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988193
Reviewed-by: John Wittrock <wittrock@google.com>
Commit-Queue: Thomas Zander <thomasalpinus@google.com>