[fxfs] Skip checksums for blobs
There was a bug in make-blob-image which resulted in data checksums
being computed for blobs (which is unnecessary, since they're already
verified).  The live writing path did not have this bug and correctly
skipped checksums.

To make this a bit simpler, I've made `skip_checksums` a handle option
rather than requiring it to be specified at the callsite (which would
require extensive plumbing, given how many write methods we have).

This significantly reduces the size of the layer file stack for initial
blob images; e.g. on my local build:

$ ls -l out/debug/root_parent_store/root_store/volumes/blob/layers
-rw-------    1 0        0            188768 Feb 06 20:35 0
-rw-------    1 0        0            311888 Feb 06 20:35 1

$ ls -l out/debug/root_parent_store/root_store/volumes/blob/layers
-rw-------    1 0        0            127208 Feb 06 20:32 0

And this also reduces the size of the fxfs.blk image by ~0.5% (e.g.
1.6MiB on a 300MiB image).

Bug: 323571978
Change-Id: Ie94af7687aa39819a013063dd6c6a196fe5439cf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987253
Commit-Queue: James Sullivan <jfsulliv@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>