[Scudo] Add custom scudo config
Now Fuchsia uses a custom config defined in Fuchsia repo instead of
relying on the config defined upstream. The new config is the same as
the one used previously, except that `RegionSizeLog` was reduced to 27
bit on RISC-V. This is a workaround to make Scudo cope with the reduced
address space on RISC-V (particularly in restricted-mode, e.g. in
Starnix).

Bug: b/321064019
Change-Id: Iab4f0945667d0d18fefac8a1e9383b59aad8d8f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/986233
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Sergey Ulanov <sergeyu@google.com>
Reviewed-by: Fabio D'Urso <fdurso@google.com>