[fxtest][debug] Allow status when in debug mode
zxdb will now not mix the outputs from debugging actions and fx test.
The result of this is the user will see the fx test status output as
normal until the debugger pops up because of some test failure. When
this happens the status output will be hidden and not mix with the
debugger's output. When finished debugging the test failure, the status
output will be seen again for the rest of the test suite (or until
another failure brings zxdb back to the foreground).

A short demonstration can be seen at
asciinema.googleplex.com/5652956951085056

Change-Id: If492ad8cf4a6b6296fce1ccc2abf2fff45942674
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987972
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Jacob Rutherford <jruthe@google.com>