[bazel] add fx bazel-refresh-compdb command
Adds an fx command which updates the compile_commands.json file with all
of the deps of a given bazel label. This command is a temporary solution
which gives us linting/error reporting for vscode.

The tool can only b called after the bazel sdk is generated or it will fail.

The code will attempt to convert what bazel thinks the paths should be
to something that maps to the fuchsia tree.

In the future this will be built into fx set but we cannot do that until
we can call bazel commands without first generating the bazel sdk.

To run the command:
fx refresh-bazel-compdb --label //build/bazel/examples/hello_cpp:pkg

BUG: b/323245932

Change-Id: I94e0f97b6b4930998a77077c090f2edede8ccab4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987966
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>