[blobfs] Use scheduler SDK library to set roles
Switch to using the scheduler SDK library to set profiles on threads
instead of connecting to the ProfileProvider directly. This will make it
much easier to migrate to the new OOT scheduler API (the RoleManager) in
the future.

Bug: 323262398
Change-Id: Idb83ec5ecebb2c118848c0f9ea2bad414ae8e1f7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/988783
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>
Reviewed-by: Chris Drouillard <cdrllrd@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>