[assembly] Size aggregator always emits all outputs
We want to be able to collect and upload size results on builds that
only run a product size checker. This means that the "budgets" and
"verbose_output" would be empty. This CL updates the aggregator rule to
generate empty files when the inputs do not contain these files.

Bug: b/319519133
Change-Id: I62f288e3a94da6ae8ebf1a7761538114be786c39
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991363
Fuchsia-Auto-Submit: Aidan Wolter <awolter@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>