[realm_builder_server] Fix ordering/timing issue
Before this CL, the tests would drop the runner server end.  A watcher
would look for peer closed on the client end and clean up which, if it
were to happen in time, would cause tests to fail.  I had a change to
fuchsia-async which caused this to trip.

This changes it so that we don't close the server end which stops the
tests from failing in that way.

Change-Id: Id9007c7ca8c587775ffcf14573251267de399f5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/991752
Reviewed-by: Claire Gonyeo <cgonyeo@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Chris Suter <csuter@google.com>