[memory_monitor] Find Starnix kernel cutoff at runtime
The cutoff address between the restricted space and the Starnix kernel
depends on the architecture. This change relies on the fact that shared
processes, as used by Starnix, have two root VMARs, one for the
restricted space and one for the Starnix kernel, to determine this
cutoff at runtime, instead of having it hardcoded.

FIXED=319838618

Change-Id: Ic09c18f29697266be6b245b72156f0174fe00370
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/981333
Reviewed-by: Fabio D'Urso <fdurso@google.com>
Commit-Queue: Étienne J. Membrives <etiennej@google.com>
Reviewed-by: Dave Moore <davemoore@google.com>