[check-licenses] Skip zircon/prebuilt
It was deprecated in fxr/735867 by adding a .gitignore entry, so
check-licenses can safely ignore it.

That CL mentions that the contents of that directory may still exist
on developer workstations, so a skip entry is needed in order to
ensure folks can run the tool locally.

Bug: b/325298549
Change-Id: I06d07644a85ace8afbec4847cad9bd2e3947fa02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/972996
Reviewed-by: Yaar Schnitman <yaar@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>