[build][riscv64] Introduce `riscv64_enable_vector`
This change introduces a GN arg to gate whether we enable RISC-V vector
instructions. We have a use-case in which we'd want to disable this and
are also currently investigating a bug in the case in which we intend to
have it enabled; this can be set to `false` locally in both of those
cases.

Bug: 42075221
Change-Id: Ifea9f81bc18fc7d7622370745e70ce6d4ad507d0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/987904
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: James Robinson <jamesr@google.com>